Rename package to hipertexto
Only the command name should be ht, the package name, from which users will install our app, should be "hipertexto".
