+++ hipertexto/images.py
