+++ hipertexto/utils.py
